organized files in animate package
