+++ app/src/main/java/com/davidfz/animfresque/ui/animate/ui_state/AnimationUiState.kt
@@ -1,4 +1,4 @@
-package com.davidfz.animfresque.ui.animate
+package com.davidfz.animfresque.ui.animate.ui_state
 
 data class AnimationUiState(
     val animationPhases: List<AnimationPhaseUiState> = listOf(),
